Separate files `container, exceptions, provider`
